refactor: 클래스 명칭을 Handler에서 Servlet으로 수정
