+++ tomcat/src/main/java/org/apache/coyote/http11/servlet/Servlet.java
@@ -1,10 +1,10 @@
-package org.apache.coyote.http11.handler;
+package org.apache.coyote.http11.servlet;
 
 import java.io.IOException;
 import org.apache.coyote.http11.common.request.HttpRequest;
 import org.apache.coyote.http11.common.response.HttpResponse;
 
-public interface Handler {
+public interface Servlet {
 
     HttpResponse handle(HttpRequest request) throws IOException;
 }
